refactor: move matrix tooling into infx package
Use infx.matrix as the canonical home for generation and validation. Keep
legacy scripts and imports compatible, materialize package code from each
historical revision, and cover import and replay boundaries with behavior tests.

将矩阵生成和校验迁移到 infx.matrix，统一模块导入。保留旧脚本和导入路径，
从每个历史修订提取对应的包代码，并通过行为测试验证导入兼容性和历史重放。
