+++ .github/workflows/test-changelog-gate.yml
@@ -46,6 +46,7 @@ on:
       - "utils/changelog_gate_tests/conftest.py"
       - "utils/merge_with_reuse.sh"
       - "utils/matrix_logic/**"
+      - "infx/**"
 
 permissions:
   contents: read
