+++ infx/matrix/__init__.py
@@ -1,0 +1,1 @@
+"""Benchmark configuration schemas and matrix generation."""
